requirements.txt
openai
PyYAML
beautifulsoup4
crawl4ai
